Move rule tests to commonrules folder
